Add first working config
This is a basic PoC commit to show it working. Lots of clean up and
refactoring to come but for now it can read in the config file and print
those METARs in the correct color with the format:
ICAO Station • Flight Category • METAR Timestamp

This will likely change going forward wanted to get a basic checkpoint
in.