Add proper config and command line parsing
This adds Cobra and Viper for proper command line parsing. Its pretty
basic right now but at least moves towards a more standard way to run
the command. This also adds some i3-type formatting. This is a WIP and
not really usable generally but it does work with one station.